Update README.md
PROJECT DESCRIPTION
 This program defines a hierarchy of restaurant classes to control and analyze their sales. 
The Restaurant class serves as an abstract base magnificence, presenting a blueprint for 
different sorts of eating places to inherit from. It has attributes for name and each day sales, 
in conjunction with techniques for showing the name and calculating sales, which must be 
carried out by subclasses. Two concrete subclasses are described, the FineDiningRestaurant 
and FastFoodRestaurant.
FineDiningRestaurant extends Restaurant, adding a score attribute. It implements the 
abstract approach calculate_revenue, considering a tip percentage. FastFoodRestaurant 
additionally extends Restaurant, with an additional attribute to suggest if it's open 24 hours. 
Its calculate_revenue approach calculates sales based totally on the assumption of running 
hours per day. At the cease of this system, instances of each subclasses are created and 
their techniques are validated. The first-class eating restaurant's name, rating, and revenue 
over a sure period are displayed, in addition to the short-meals eating place's call, opening 
hours, and sales over a detailed variety of days